Refactor ft_calloc, ft_memmove, and ft_strncmp to update timestamps and improve code clarity

+++ ft_calloc.c
@@ -6,13 +6,13 @@
 /*   By: rimagalh <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/23 15:54:07 by rimagalh          #+#    #+#             */
-/*   Updated: 2024/10/28 18:50:01 by rimagalh         ###   ########.fr       */
+/*   Updated: 2024/11/02 21:34:11 by rimagalh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static int	valid_size(size_t nmemb, size_t size)
+static size_t	valid_size(size_t nmemb, size_t size)
 {
 	size_t	max_size;
